+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes
 
 - Breaking change: Rename `advertisement.UpnpAdvertisementListener` to `advertisement.SsdpAdvertisementListener`
 - Breaking change: Rename `search.SSDPListener` to `search.SsdpSearchListener`
+- Add `ssdp_listener.SsdpListener`, class to keep track of devices seen via SSDP advertisements and searches
 
 
 0.20.0 (2021-08-17)